Refactor FPM with its functional tests
The new approach leaves less footprint in the system and doesn't depend
on docker-compose.
Also we now have a more accurate test matrix
