+++ .circleci/config.yml
@@ -103,17 +103,10 @@ jobs:
       - checkout
       - attach_workspace:
           at: ./tmp
-      - run:
-          name: Update Docker Compose
-          command: |
-            sudo curl -L https://github.com/docker/compose/releases/download/1.22.0/docker-compose-$(uname -s)-$(uname -m) -o /usr/local/bin/docker-compose
-            sudo chmod +x /usr/local/bin/docker-compose
-      - run: docker load -i ./tmp/usabillabv_php-http.tar
       - run: docker load -i ./tmp/usabillabv_php-fpm.tar
-      - run: docker load -i ./tmp/usabillabv_php-cli.tar
-      - run: make ci-test-fpm
+      - run: make test-fpm
       - store_test_results:
-          path: ./test/test-results/
+          path: ./tmp/test-results
   scan-vulnerability:
     machine: true
     steps:
