+++ .gitignore
@@ -1,1 +1,2 @@
 /tmp
+/test/functional/web/tmp
